backend/requirements.txt
# Django and Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Other utilities
Pillow==10.1.0
python-dateutil==2.8.2